pyproject.toml
[project]
name = "tiktok-image-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.64.0",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.0",
]
